fix(core): type rare_threshold as f64, not usize
rare_threshold is now compared against a time-decay weighted sum
(f64), not an integer occurrence count. Keeping the type as usize
silently changed --remove-rare behavior: 20 commands older than 30d
accumulate only 2.0 weight and fall under the default threshold of 3,
making them deletion-eligible despite being used 20 times.

Changing the type to f64 surfaces the semantic shift in the public
API and removes the implicit as-f64 cast. Default stays 3.0 — the
same boundary, now interpreted as a weighted score. README updated
to describe weights and accept float values.

+++ src/cleaner.rs
@@ -138,7 +138,7 @@ fn rare_variants(
 
     let by_base = group_by_base_strs(weighted_counts.keys().copied());
     for (&rare_cmd, &rare_weight) in &weighted_counts {
-        if rare_weight > settings.rare_threshold as f64 {
+        if rare_weight > settings.rare_threshold {
             continue;
         }
         let base = base_command(rare_cmd);
